Refine configuration of sync rules
The config parser won't treat all options in the TagSyncher section as rules
anymore. Instead, a new option 'folders' is required that contains a whitespace
separated list of maildir folders that are to be synced. Only options with keys
matching a specified folder will be considered rules.

The new option is mandatory when calling afew in sync mode.
For each specified folder an option with a corresponding key is mandatory.